Blog rest api

+++ src/posts/views.py
@@ -7,6 +7,9 @@ from django.urls import reverse
 from django.db.models import Q,F
 from django.core.paginator import Paginator
 from django.utils import timezone
+from faker import Faker
+import logging
+import logging.config
 
 from .models import Post
 from .forms import PostForm
@@ -17,8 +20,12 @@ from comments.forms import CommentForm
 
 # Create your views here.
 
+logging.config.fileConfig(fname='logs/log.conf')
+logger = logging.getLogger('posts')
+
 def post_index(request,slug = None):
     qs_list = Post.objects.all()
+    logger.info(f'{qs_list.count()} posts listed.')
     if slug is not None:
         qs_list = Post.objects.filter(category__icontains=slug)
     query = request.GET.get("q")
@@ -29,7 +36,7 @@ def post_index(request,slug = None):
             Q(user__first_name__icontains=query) |
             Q(user__last_name__icontains=query)
         ).distinct()
-    paginator = Paginator(qs_list, 3)
+    paginator = Paginator(qs_list, 4)
     page_request_var = 'page'
     page_number = request.GET.get(page_request_var)
     page_obj = paginator.get_page(page_number)
@@ -38,21 +45,25 @@ def post_index(request,slug = None):
 
 def post_create(request):
     if not request.user.is_authenticated:
+        logger.error(f'Not Authenticated')
         return HttpResponse('You are not authenticated')
     form = PostForm(request.POST or None,request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
         instance.user = request.user
         instance.save()
-        messages.success(request,'Post created.')
+        logger.info(f'Post created by {request.user} with title {instance.title}')
+        messages.success(request,'Post created. ')
         return HttpResponseRedirect(instance.get_absolute_url())
     context = {'form':form}
     return render(request,template_name='posts/create.html',context=context)
 
 def post_list(request):
     if not request.user.is_authenticated:
+        logger.error(f'Not Authenticated')
         return HttpResponse('You are not authenticated')
     qs_list = Post.objects.filter(user=request.user)
+    logger.info(f'{request.user} listed {qs_list.count()} posts.')
     query = request.GET.get("q")
     if query:
         qs_list = qs_list.filter(
@@ -74,11 +85,18 @@ def post_details(request,slug=None):
         Http404
     qs_details = queryset[0]
     if qs_details.draft or qs_details.publish > timezone.now().date():
-        if not request.user.is_authenticated:
-            return HttpResponse('You are not authenticated')
+        Http404
+    if not request.user.is_authenticated:
+        return HttpResponse('You are not authenticated')
     if request.method == 'GET':
         queryset.update(views_count = F('views_count') + 1)
     qs_details.refresh_from_db()
+
+    category = qs_details.category
+    print(category)
+    queryset_category = Post.objects.filter(category=category).order_by("-views_count")
+    if queryset_category.count() > 3:
+        queryset_category = queryset_category[:3]
 
     initial_data = {'content_type':qs_details.get_content_type,'object_id':qs_details.id}
     form = CommentForm(request.POST or None,initial=initial_data)
@@ -106,27 +124,48 @@ def post_details(request,slug=None):
                                     content =content_data,
                                     parent = parent_obj,
         )
+        logger.info(f'{request.user} commented {new_comment.content}.')
         return HttpResponseRedirect(new_comment.content_object.get_absolute_url())
 
     comments = qs_details.comment
-    context = {'query':qs_details,'title':qs_details.title,'comments':comments,'comment_form':form}
+    context = {'query':qs_details,'title':qs_details.title,'comments':comments,'comment_form':form,'queryset_category':queryset_category}
     return render(request,template_name='posts/details.html',context=context)
 
 def post_update(request,slug=None):
+    if not request.user.is_authenticated:
+        return HttpResponse('You are not authenticated')
     qs_update = get_object_or_404(Post,slug=slug)
     form = PostForm(request.POST or None,request.FILES or None,instance=qs_update)
     if form.is_valid():
         instance = form.save(commit=False)
         instance.save()
+        logger.info(f'{request.user} updated {instance.title}.')
         return HttpResponseRedirect(instance.get_absolute_url())
     context = {'query':qs_update,'title':qs_update.title,'form':form}
     return render(request,template_name='posts/edit.html',context=context)
 
 def post_delete(request,slug=None):
-    qs_delete = get_object_or_404(Post,slug=slug)
-    qs_delete.delete()
-    messages.success(request,'Succesfully deleted.')
-    return HttpResponseRedirect(reverse('posts:postlist'))
+    try:
+        qs_delete = get_object_or_404(Post,slug=slug)
+    except:
+        raise Http404('This is not present')
+
+    if not request.user.is_authenticated:
+        return HttpResponse('You are not authenticated')
+
+    if qs_delete.user != request.user:
+        response = HttpResponse("you dont have permission to view this.")
+        response.status_code = 403
+        return response
+
+    if request.method == 'POST':
+        qs_delete.delete()
+        logger.info(f'{request.user} deleted {qs_delete}.')
+        messages.success(request,'Post has been deleted')
+        return HttpResponseRedirect(reverse('posts:postlist'))
+
+    context = {'qs_delete':qs_delete}
+    return render(request,'posts/post_delete.html',context=context)
 
 def contact(request):
     return render(request,template_name='posts/contact.html')
